add main logic refactor.
