+++ .eslintrc.json
@@ -11,10 +11,6 @@
     "ecmaVersion": "latest"
   },
   "rules": {
-    "class-methods-use-this": 0,
-    "no-param-reassign": 0,
-    "no-restricted-syntax": [0, "BinaryExpression[operator='in']"],
-    "no-await-in-loop": 0,
     "no-console": 0
   }
 }
