+++ src/templateList.js
@@ -1,0 +1,14 @@
+const templates = [
+  {
+    value: "minimal-react",
+    title: "minimal react app",
+    description: "Minimal React App Set Up.",
+  },
+  {
+    value: "vanilla-js",
+    title: "vanilla js app",
+    description: "Vanilla JS App Set Up.",
+  },
+];
+
+export default templates;
